refactor: Cleanup
chore: update webui build output
